requirements.txt
pycryptodome
pwntools
